+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ## 2.6 branch
 ### 2.6.4
 * now you can show banners and photos as list or as grid;
-* now you can download banners.
+* now you can download banners and photos.
 
 ### 2.6.3
 * fixed bug ordering posts and pages.